change file of javacript of map
