--- scss/styles.scss
@@ -1,3 +1,0 @@
-canvas{
-    margin:calc(50vh-200px) calc(50vw - 375px)
-}
